session_15/pyproject.toml
[project]
name = "session-15"
version = "0.1.0"
description = ""
authors = [
    {name = "NelsonZarate",email = "nelson_zarate@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)"
]

[tool.poetry]
packages = [{include = "djproject"}]


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.10.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "djproject.settings"